projects/interview_data_analysis/lda/requirements.txt
# 機械学習・データ分析
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# 日本語形態素解析
mecab-python3>=1.0.0

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0
japanize-matplotlib>=1.1.0
wordcloud>=1.9.0

# その他
scipy>=1.10.0
